Fix stop after toggle bug
