+++ src/main/kotlin/com/abysl/harryplotter/windows/HarryPlotterSettingsWindow.kt
@@ -1,0 +1,8 @@
+import com.abysl.harryplotter.view.settings.HarryPlotterSettingsView
+import com.abysl.harryplotter.windows.Window
+
+class HarryPlotterSettingsWindow: Window<HarryPlotterSettingsView>(){
+    fun show(){
+        create("Harry Plotter Settings", "fxml/settings/HarryPlotterSettings.fxml")
+    }
+}
