+++ src/main/resources/com/abysl/harryplotter/fxml/settings/HarryPlotterSettings.fxml
@@ -1,0 +1,43 @@
+<?xml version="1.0" encoding="UTF-8"?>
+
+<?import javafx.geometry.Insets?>
+<?import javafx.scene.control.Button?>
+<?import javafx.scene.layout.HBox?>
+<?import javafx.scene.layout.VBox?>
+
+<!--
+  ~     Copyright (c) 2021 Andrew Bueide
+  ~     
+  ~     This file is part of Harry Plotter.
+  ~
+  ~     Harry Plotter is free software: you can redistribute it and/or modify
+  ~     it under the terms of the GNU General Public License as published by
+  ~     the Free Software Foundation, either version 3 of the License, or
+  ~     (at your option) any later version.
+  ~
+  ~     Harry Plotter is distributed in the hope that it will be useful,
+  ~     but WITHOUT ANY WARRANTY; without even the implied warranty of
+  ~     MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+  ~     GNU General Public License for more details.
+  ~
+  ~     You should have received a copy of the GNU General Public License
+  ~     along with Harry Plotter.  If not, see <https://www.gnu.org/licenses/>.
+  -->
+
+<VBox alignment="CENTER" prefHeight="400.0" prefWidth="600.0" spacing="10.0" xmlns="http://javafx.com/javafx/16" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.abysl.harryplotter.view.settings.HarryPlotterSettingsView">
+   <padding>
+      <Insets bottom="10.0" left="10.0" right="10.0" top="10.0" />
+   </padding>
+   <children>
+      <VBox alignment="CENTER" prefHeight="200.0" prefWidth="100.0" VBox.vgrow="ALWAYS">
+         <children>
+            <Button mnemonicParsing="false" onAction="#onReset" text="Reset Settings" />
+         </children>
+      </VBox>
+      <HBox alignment="BOTTOM_RIGHT">
+         <children>
+            <Button fx:id="close" mnemonicParsing="false" onAction="#onClose" text="Close" />
+         </children>
+      </HBox>
+   </children>
+</VBox>
