+++ src/main/resources/com/abysl/harryplotter/version.txt
@@ -1,1 +1,1 @@
-1.1.8
+1.1.7
